Fix reflect:equals() returns FALSE for tuples
Suppose there are two tuples x and y which are equals. Reflect stdlib API
reflect:equals() does not return TRUE even they are same. This was due to fact that;
Tuple type has not been considered. With this commit added support for the Tuples.

This commit resolves https://github.com/ballerina-platform/ballerina-lang/issues/10481